Abstracted out logic to usePane hook
