+++ lib/components/Pane/index.js
@@ -1,16 +1,15 @@
 import React, { useRef } from "react";
+import classnames from "classnames";
 import PropTypes from "prop-types";
 import { Close } from "@bigbinary/neeto-icons";
-import { useHotkeys } from "react-hotkeys-hook";
-import classnames from "classnames";
 import { CSSTransition } from "react-transition-group";
 
-import { Portal, Backdrop } from "../../atoms";
-import { useOnClickOutside } from "../../utils";
-import Button from "../Button";
-import Header from "./Header";
 import Body from "./Body";
+import Header from "./Header";
 import Footer from "./Footer";
+import Button from "../Button";
+import { usePane } from "../../hooks/usePane";
+import { Portal, Backdrop } from "../../atoms";
 
 const noop = () => {};
 
@@ -31,22 +30,14 @@ const Pane = ({
   const paneWrapper = useRef();
   const backdropRef = useRef();
 
-  useOnClickOutside(
+  usePane({
     paneWrapper,
     backdropRef,
-    closeOnOutsideClick ? onClose : noop
-  );
-
-  const onCloseRef = useRef();
-  onCloseRef.current = onClose;
-
-  useHotkeys(
-    "esc",
-    () => {
-      closeOnEsc ? onCloseRef.current() : noop();
-    },
-    [onCloseRef]
-  );
+    closeOnOutsideClick,
+    closeOnEsc,
+    onClose,
+    isOpen,
+  });
 
   return (
     <Portal rootId="neeto-ui-portal">
