+++ lib/hooks/useModal.js
@@ -2,7 +2,11 @@ import { useRef, useCallback, useEffect } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 
 import { manager } from "../components/Modal/ModalManager";
-import { focusElementsInModal, useOnClickOutside, noop } from "../utils";
+import {
+  renderFocusOnFocusableElements,
+  useOnClickOutside,
+  noop,
+} from "../utils";
 
 export const useModal = ({
   isOpen,
@@ -31,8 +35,8 @@ export const useModal = ({
   const focusRequiredElementInModal = () => {
     if (initialFocusRef?.current) {
       initialFocusRef.current.focus();
-      focusElementsInModal(modalWrapper, false);
-    } else focusElementsInModal(modalWrapper, true);
+      renderFocusOnFocusableElements(modalWrapper, false);
+    } else renderFocusOnFocusableElements(modalWrapper, true);
   };
 
   const handleModalClose = useCallback(() => {
