+++ lib/utils.js
@@ -77,7 +77,10 @@ export class UniqueArray {
   }
 }
 
-export const focusElementsInModal = (ref, shouldFocusFirstFocusableElement) => {
+export const renderFocusOnFocusableElements = (
+  ref,
+  shouldFocusFirstFocusableElement
+) => {
   const focusableElements =
     'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])';
   const firstFocusableElement = ref?.current?.querySelectorAll(
